Added authorization for records

+++ controllers/record.controller.js
@@ -5,8 +5,14 @@ const RecordService = require("../services/record/record.service.js");
 const asyncHandler = require("../middleware/errors/asyncHandler.js");
 
 const show = asyncHandler(async (req, res) => {
+  const user = req.user;
   const name = req.query.name || "";
   const { workspaceId, collectionId, recordId } = req.params;
+
+  const workspace = WorkspaceService.requireViewableWorkspace({
+    workspaceId,
+    user,
+  });
 
   const details = await RecordService.getDetails({
     workspaceId,
@@ -15,18 +21,22 @@ const show = asyncHandler(async (req, res) => {
   });
 
   res.render("record-details", {
-    workspace: details.workspace,
+    workspace,
     collection: details.collection,
     record: details.record,
-    user: req.user,
+    user,
     name,
   });
 });
 
 const newRecord = asyncHandler(async (req, res) => {
+  const user = req.user;
   const { workspaceId, collectionId } = req.params;
 
-  const workspace = await WorkspaceService.requireWorkspace({ workspaceId });
+  const workspace = await WorkspaceService.requireOwnerWorkspace({
+    workspaceId,
+    user,
+  });
 
   const collection = await CollectionService.requireCollection({
     workspaceId,
@@ -34,7 +44,7 @@ const newRecord = asyncHandler(async (req, res) => {
   });
 
   res.render("record-form", {
-    user: req.user,
+    user,
     workspace,
     collection,
     record: null,
@@ -44,9 +54,15 @@ const newRecord = asyncHandler(async (req, res) => {
 });
 
 const create = asyncHandler(async (req, res) => {
+  const user = req.user;
   const { workspaceId, collectionId } = req.params;
   const { title, description } = req.body;
   const { files } = req;
+
+  await WorkspaceService.requireOwnerWorkspace({
+    workspaceId,
+    user,
+  });
 
   await RecordService.create({
     workspaceId,
@@ -60,7 +76,13 @@ const create = asyncHandler(async (req, res) => {
 });
 
 const edit = asyncHandler(async (req, res) => {
+  const user = req.user;
   const { workspaceId, collectionId, recordId } = req.params;
+
+  const workspace = await WorkspaceService.requireOwnerWorkspace({
+    workspaceId,
+    user,
+  });
 
   const details = await RecordService.getDetails({
     workspaceId,
@@ -71,7 +93,7 @@ const edit = asyncHandler(async (req, res) => {
   res.render("record-form", {
     user: req.user,
     collection: details.collection,
-    workspace: details.workspace,
+    workspace,
     record: details.record,
     files: null,
     errors: [],
@@ -79,9 +101,15 @@ const edit = asyncHandler(async (req, res) => {
 });
 
 const update = asyncHandler(async (req, res) => {
+  const user = req.user;
   const { workspaceId, collectionId, recordId } = req.params;
   const { title, description, deletedFiles } = req.body;
   const { files } = req;
+
+  await WorkspaceService.requireOwnerWorkspace({
+    workspaceId,
+    user,
+  });
 
   await RecordService.update({
     title,
@@ -97,7 +125,13 @@ const update = asyncHandler(async (req, res) => {
 });
 
 const remove = asyncHandler(async (req, res) => {
+  const user = req.user;
   const { workspaceId, collectionId, recordId } = req.params;
+
+  await WorkspaceService.requireOwnerWorkspace({
+    workspaceId,
+    user,
+  });
 
   const recordRemoved = await RecordService.remove({
     workspaceId,
